refactor(components/react/hooks): rename folder
