+++ components/react/hooks/src/index.js
@@ -33,4 +33,4 @@ export {default as useSwipe} from './useSwipe/index.js'
 export {default as useToggle} from './useToggle/index.js'
 export {default as useIsomorphicLayoutEffect} from './useIsomorphicLayoutEffect/index.js'
 export {default as useCallbackRef} from './useCallbackRef/index.js'
-export {default as usePreCalculatedMediaQuery} from './usePrecalculatedMediaQuery/index.js'
+export {default as usePreCalculatedMediaQuery} from './usePreCalculatedMediaQuery/index.js'
